Added function to write timeframes in format of SystemC simulation
This function writes the timeframe in the format used by the SystemC
SAMPA simulation of the master project of H.R. Olsen. It might be
combined with the already existing function WriteTimeframe.

Long term plan: either define a robust format for the ascii file or
connect the ChannelMerger directly with the dataGenerator class of
the simulation.