GH Fix #1: Corrected treatment of pedestal value and baselineshift
Baselineshift can now exceed the value of the pedestal. The implementation
though appears to be a bit awkard seen retrospective. Probably its better
to store the pedestal values with original values rather the adjusting them
by the baselineshift. The baselineshift needs to be taken into account in
many places to correct for this adjustment. It would be clearer with the
unchanged pedestal values and the baselineshift treatment directly where
it is needed.